# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Users API
  description: |
    API RESTful para gerenciamento de usuários desenvolvida com Spring Boot 3.
    
    Esta API permite:
    - ✅ Listar todos os usuários
    - ✅ Criar novos usuários
    - ✅ Buscar usuários por ID
    - ✅ Atualizar usuários existentes
    - ✅ Remover usuários
    
    ### Tecnologias utilizadas:
    - **Java 17**
    - **Spring Boot 3**
    - **Spring Web**
    - **Spring Data JPA**
    - **Spring Validation**
    - **H2 Database (in-memory)**
    - **Lombok**
    - **SpringDoc OpenAPI**
    
    ### Banco de dados:
    A aplicação utiliza H2 Database em memória para desenvolvimento.
    Os dados são perdidos a cada reinicialização da aplicação.
    
    ### Console do H2:
    Acesse o console do banco em: [/h2-console](http://localhost:8080/h2-console)
    - **URL**: jdbc:h2:mem:testdb
    - **Usuário**: sa
    - **Senha**: (vazio)
  version: 1.0.0
  contact:
    name: Equipe de Desenvolvimento
    email: dev@example.com
    url: https://example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento Local
  - url: https://api.example.com
    description: Servidor de Produção (Exemplo)

tags:
  - name: Users
    description: API para gerenciamento de usuários

paths:
  /users:
    get:
      tags:
        - Users
      summary: Lista todos os usuários
      description: Retorna uma lista com todos os usuários cadastrados no sistema
      operationId: getAllUsers
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              examples:
                usuarios_exemplo:
                  summary: Exemplo de lista de usuários
                  value:
                    - id: 1
                      name: "João Silva"
                      email: "joao@email.com"
                      age: 30
                      phone: "(11) 99999-9999"
                      createdAt: "2024-01-15T10:00:00"
                      updatedAt: "2024-01-15T10:00:00"
                    - id: 2
                      name: "Maria Santos"
                      email: "maria@email.com"
                      age: 25
                      phone: "(11) 88888-8888"
                      createdAt: "2024-01-15T11:00:00"
                      updatedAt: "2024-01-15T11:00:00"

    post:
      tags:
        - Users
      summary: Cria novo usuário
      description: Cadastra um novo usuário no sistema
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              usuario_completo:
                summary: Usuário com todos os campos
                value:
                  name: "Pedro Oliveira"
                  email: "pedro@email.com"
                  age: 28
                  phone: "(11) 77777-7777"
              usuario_basico:
                summary: Usuário apenas com campos obrigatórios
                value:
                  name: "Ana Costa"
                  email: "ana@email.com"
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                usuario_criado:
                  summary: Usuário criado
                  value:
                    id: 3
                    name: "Pedro Oliveira"
                    email: "pedro@email.com"
                    age: 28
                    phone: "(11) 77777-7777"
                    createdAt: "2024-01-15T12:00:00"
                    updatedAt: "2024-01-15T12:00:00"
        '400':
          description: Dados inválidos ou email já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_duplicado:
                  summary: Email já existe
                  value:
                    timestamp: "2024-01-15T12:00:00"
                    status: 400
                    error: "Bad Request"
                    message: "Email já está em uso: pedro@email.com"
                validacao_erro:
                  summary: Erro de validação
                  value:
                    timestamp: "2024-01-15T12:00:00"
                    status: 400
                    error: "Validation Failed"
                    message: "Dados inválidos fornecidos"
                    errors:
                      name: "Nome é obrigatório"
                      email: "Email deve ter um formato válido"

  /users/{id}:
    get:
      tags:
        - Users
      summary: Busca usuário por ID
      description: Retorna os detalhes de um usuário específico
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                usuario_encontrado:
                  summary: Usuário encontrado
                  value:
                    id: 1
                    name: "João Silva"
                    email: "joao@email.com"
                    age: 30
                    phone: "(11) 99999-9999"
                    createdAt: "2024-01-15T10:00:00"
                    updatedAt: "2024-01-15T10:00:00"
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                usuario_nao_encontrado:
                  summary: Usuário não encontrado
                  value:
                    timestamp: "2024-01-15T12:00:00"
                    status: 404
                    error: "Not Found"
                    message: "Usuário não encontrado com ID: 999"

    put:
      tags:
        - Users
      summary: Atualiza usuário existente
      description: Atualiza todos os dados de um usuário específico
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            examples:
              atualizacao_completa:
                summary: Atualização completa
                value:
                  name: "João Silva Santos"
                  email: "joao.santos@email.com"
                  age: 31
                  phone: "(11) 99999-0000"
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                usuario_atualizado:
                  summary: Usuário atualizado
                  value:
                    id: 1
                    name: "João Silva Santos"
                    email: "joao.santos@email.com"
                    age: 31
                    phone: "(11) 99999-0000"
                    createdAt: "2024-01-15T10:00:00"
                    updatedAt: "2024-01-15T13:00:00"
        '400':
          description: Dados inválidos ou email já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Remove usuário
      description: Remove um usuário do sistema
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Usuário removido com sucesso
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                usuario_nao_encontrado:
                  summary: Usuário não encontrado para remoção
                  value:
                    timestamp: "2024-01-15T12:00:00"
                    status: 404
                    error: "Not Found"
                    message: "Usuário não encontrado com ID: 999"

components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Nome do usuário
          minLength: 2
          maxLength: 100
          example: "João Silva"
        email:
          type: string
          format: email
          description: Email do usuário
          maxLength: 150
          example: "joao@email.com"
        age:
          type: integer
          description: Idade do usuário (opcional)
          minimum: 0
          maximum: 150
          example: 30
        phone:
          type: string
          description: Telefone do usuário (opcional)
          maxLength: 20
          example: "(11) 99999-9999"

    UpdateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Nome do usuário
          minLength: 2
          maxLength: 100
          example: "João Silva Santos"
        email:
          type: string
          format: email
          description: Email do usuário
          maxLength: 150
          example: "joao.santos@email.com"
        age:
          type: integer
          description: Idade do usuário (opcional)
          minimum: 0
          maximum: 150
          example: 31
        phone:
          type: string
          description: Telefone do usuário (opcional)
          maxLength: 20
          example: "(11) 99999-0000"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único do usuário
          example: 1
        name:
          type: string
          description: Nome do usuário
          example: "João Silva"
        email:
          type: string
          format: email
          description: Email do usuário
          example: "joao@email.com"
        age:
          type: integer
          description: Idade do usuário
          example: 30
        phone:
          type: string
          description: Telefone do usuário
          example: "(11) 99999-9999"
        createdAt:
          type: string
          format: date-time
          description: Data de criação do usuário
          example: "2024-01-15T10:00:00"
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização
          example: "2024-01-15T10:00:00"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp do erro
          example: "2024-01-15T12:00:00"
        status:
          type: integer
          description: Código de status HTTP
          example: 400
        error:
          type: string
          description: Tipo do erro
          example: "Bad Request"
        message:
          type: string
          description: Mensagem do erro
          example: "Email já está em uso"
        errors:
          type: object
          description: Detalhes dos erros de validação (opcional)
          additionalProperties:
            type: string
          example:
            name: "Nome é obrigatório"
            email: "Email deve ter um formato válido"